feat: move form security to separate class.
